+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 
 **JS:**
 
-- `Improved` #733 Selected 增加超过最多可选值时提示接口。
+- `Improved` #733 Selected 增加超过最多可选值时提示接口；
+- `Improved` #749 Selected 允许添加一个空的 `<option>` 实现不自动选择第一项。
 
 ### 2015.11 W2